Added functionality for partial time values

+++ ChoirNote.java
@@ -2,10 +2,10 @@ public class ChoirNote extends Note
 {
 	//TODO: add support for dotted and smaller valued time values
 	//ex. 0.5 duration as en eighth note in 4/4
-	int duration;
+	float duration;
 	int octave;
 
-	public ChoirNote(String note, int octave, int duration)
+	public ChoirNote(String note, int octave, float duration)
 	{
 		super(note);
 		this.duration = duration;
